+++ smile_id_core/Signature.py
@@ -30,5 +30,5 @@ class Signature:
             "timestamp": timestamp,
         }
 
-    def confirm_signature(self, timestamp: str, msg_signature):
+    def confirm_signature(self, timestamp: str, msg_signature: str) -> bool:
         return self.generate_signature(timestamp)["signature"] == msg_signature